Move helper functions to speciesTabsWrapperHelpers and account for margins
